+++ src/main/java/me/pajic/tib/TIBUtil.java
@@ -2,7 +2,9 @@ package me.pajic.tib;
 
 import net.minecraft.client.gui.screens.inventory.CreativeModeInventoryScreen;
 import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
+import net.minecraft.tags.ItemTags;
 import net.minecraft.world.Container;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
@@ -20,7 +22,10 @@ public class TIBUtil {
     public static void addArrowsFromContainer(Container container, MutableTriple<Integer, Integer, Boolean> arrows) {
         for (int i = 0; i < container.getContainerSize(); i++) {
             ItemStack is = container.getItem(i);
-            if (is.getItem() == Items.ARROW || is.getItem() == Items.SPECTRAL_ARROW || is.getItem() == Items.TIPPED_ARROW) {
+            //? if 1.21.1
+            if (BuiltInRegistries.ITEM.getTag(ItemTags.ARROWS).orElseThrow().contains(is.getItemHolder())) {
+            //? if 1.21.8
+            /*if (BuiltInRegistries.ITEM.getOrThrow(ItemTags.ARROWS).contains(is.getItemHolder())) {*/
                 arrows.setLeft(arrows.getLeft() + is.getCount());
                 if (is.getItem() == Items.ARROW) arrows.setRight(true);
                 else arrows.setMiddle(arrows.getMiddle() + is.getCount());